Former-commit-id: adaed1cd8cfc79273f274e3b2cb9d77a008d81c2

+++ Sources/Internal/Render/RenderHelper.cpp
@@ -324,7 +324,7 @@ void RenderHelper::DrawLines(const Vector<float32>& linePoints, UniqueHandle ren
     RenderManager::Instance()->SetRenderEffect(RenderSystem2D::FLAT_COLOR);
     RenderManager::Instance()->SetRenderData(renderDataObject);
 
-    RenderManager::Instance()->DrawArrays(PRIMITIVETYPE_LINELIST, 0, linePoints.size() / 2);
+    RenderManager::Instance()->DrawArrays(PRIMITIVETYPE_LINELIST, 0, static_cast<int32>(linePoints.size() / 2));
 }
 
 
@@ -934,7 +934,7 @@ void RenderHelper::DrawCornerBox(const AABBox3 & bbox, float32 lineWidth, Unique
 		
 		//draw
 		
-		int32 size = triangleIndices.size()/3;
+		int32 size = static_cast<int32>(triangleIndices.size()/3);
 		for (int i = 0; i < size; i++)
 		{
 			Vector3 p1 = points[triangleIndices[i]] + center;
@@ -1000,7 +1000,7 @@ void RenderHelper::DrawCornerBox(const AABBox3 & bbox, float32 lineWidth, Unique
 
 		//fill
 
-		int32 size = triangleIndices.size()/3;
+		int32 size = static_cast<int32>(triangleIndices.size()/3);
 		for (int i = 0; i < size; i++)
 		{
 			Vector3 p1 = points[triangleIndices[i]] + center;
